Add button to show logs
- Move client code to /src/lib/services-manager/clients/
